+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 * Detect user input for CommandBodyBox and CommentBox and create a save button in such events.]
 * Add a button beside the command body to copy the command to clipboard.
 * Add a button to create a command for the current selected collection.
+* Add extract_options() for command model class to extract options(with -- or -) as a list.
+* Add unit tests for extract_options() and get_template_variables() for the command model class.
 
 ### CHANGED
 * Use static images for some buttons